requirements.txt
asn1crypto==0.23.0
backports.ssl-match-hostname==3.5.0.1
cffi==1.9.1
chardet==2.3.0
chrome-gnome-shell==0.0.0
cryptography==2.0.2
dj-database-url==0.5.0
Django==1.10
django-crispy-forms==1.7.0
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==1.0.4
django-registration-redux==2.2
enum34==1.1.6
gunicorn==19.7.1
idna==2.5
iniparse==0.4
ipaddress==1.0.16
Markdown==2.6.11
markdown2==2.3.5
ply==3.9
pycparser==2.14
pycurl==7.43.0
pygobject==3.24.1
pygpgme==0.3
pyliblzma==0.5.3
pyOpenSSL==16.2.0
PySocks==1.6.7
pytz==2018.3
pyxattr==0.5.3
requests==2.13.0
rpm-python==4.13.0.2
six==1.10.0
urlgrabber==3.10.1
urllib3==1.20
virtualenv==15.1.0
youtube-dl==2018.1.21
yum-metadata-parser==1.1.4
